moved AbstractWorkflowBasedRunConfiguration to de.uka.ipd.sdq.workflow.launchconfig.core
