+++ bundles/de.uka.ipd.sdq.workflow.launchconfig.core/META-INF/MANIFEST.MF
@@ -3,6 +3,9 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Run Config Core
 Bundle-SymbolicName: de.uka.ipd.sdq.workflow.launchconfig.core
 Bundle-Version: 5.2.1.qualifier
+Export-Package: de.uka.ipd.sdq.workflow.launchconfig.core
+Require-Bundle: org.eclipse.debug.core;bundle-version="3.20.0",
+ de.uka.ipd.sdq.workflow;bundle-version="5.2.1"
 Bundle-ActivationPolicy: lazy
 Bundle-Vendor: sdq.ipd.uka.de
 Bundle-RequiredExecutionEnvironment: JavaSE-17
